Updates an error message to be much more prominent (#182)
It now looks something like this:
![alt text](https://i.imgur.com/dvKtdVG.png1)

I realised whilst looking at this that we probably shouldnt have that error message in bolt itself because it assumes the existance of changesets. Im happy to leave for now though until `bolt` has a blessed way of handling this (it'll be a minor text change).