Implement help command and modify --help behaviour
--help now only displays help information if ran with no commands, i.e.
`bolt --help`.
If a command is passed, e.g. bolt <command> --help`. the `--help` flag will be
passed through to that command and not display help text.
This fixes running npm scripts that have implemented their own flag,
e.g. `yarn run my-script-that-has-a-help-flag --help`

In the future we will implement help information for subcommands that
may handle `bolt <command> --help` as well as `bolt help <subcommand>`.